+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/RoomListRouter.kt
@@ -180,6 +180,9 @@ class RoomListRouter(
                     viewModelContext = viewModelContext.childContext(componentContext),
                     onCloseProfile = ::onCloseProfile,
                     onOpenAvatarCutter = onOpenAvatarCutter,
+                    onShowAccountSetup = ::onShowAccountSetup,
+                    onRemoveAccount = onRemoveAccount,
+                    onCreateNewAccount = onCreateNewAccount,
                 )
             )