+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/settings/ProfileViewModelTest.kt
@@ -331,6 +331,9 @@ class ProfileViewModelTest {
             ),
             onCloseProfile = mock(),
             onOpenAvatarCutter = mock(),
+            onShowAccountSetup = mock(),
+            onRemoveAccount = mock(),
+            onCreateNewAccount = mock(),
         )
     }
 }